MOD: Finished with package and source code changes aggregation
Finished implementation of the pulling meta information about
versions as well as detecting difference (in commits) between
installed and consequtive versions for a NPM package and Github
repository.

Next stop is to create an overview of what is currently
goint on version-wise as well as code age and summary of
the newer changes.