MOD: Reorganized pyproject dependencies and updated docs
Updated docs to clearly communicate what is possible and
what is not possible with the current implementation,
especially around PyPI ecosystem.

Reorganized dependencies in pyproject to have two
categories of optional dependencies: one is dev and
another one is docs to keep dependencies for documentation
and active development respectively.

GH-4